Se realizaron los test de LibroController, con los que se detactarion errores en el amnejo de los errores y los mensajes de htts status (se corrigieron)
