--- biblioteca/src/main/java/Controllers/LibroController.java
@@ -1,43 +1,0 @@
-package Controllers;
-
-import Modelos.Libro;
-import Services.LibroService;
-
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("api/libros")
-public class LibroController {
-    private final LibroService libroService;
-
-    public LibroController(LibroService libroService) {
-        this.libroService = libroService;
-    }
-
-    @GetMapping
-    public List<Libro> obtenerTodos() {
-        return libroService.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public Libro obtenerPorId(@PathVariable Long id) {
-        return libroService.findById(id);
-    }
-
-    @PostMapping
-    public Libro crear(@RequestBody Libro libro) {
-        return libroService.save(libro);
-    }
-
-    @PutMapping("/{id}")
-    public Libro actualizar(@PathVariable Long id, @RequestBody Libro libro) {
-        return libroService.update(id, libro);
-    }
-
-    @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable Long id) {
-        libroService.delete(id);
-    }
-}
