--- biblioteca/src/main/java/Controllers/UsuarioController.java
@@ -1,43 +1,0 @@
-package Controllers;
-
-import Modelos.Usuario;
-import Services.UsuarioService;
-
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("api/usuarios")
-public class UsuarioController {
-    private final UsuarioService usuarioService;
-
-    public UsuarioController(UsuarioService usuarioService) {
-        this.usuarioService = usuarioService;
-    }
-
-    @GetMapping
-    public List<Usuario> obtenerTodos() {
-        return usuarioService.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public Usuario obtenerPorId(@PathVariable Long id) {
-        return usuarioService.findById(id);
-    }
-
-    @PostMapping
-    public Usuario crear(@RequestBody Usuario usuario) {
-        return usuarioService.save(usuario);
-    }
-
-    @PutMapping("/{id}")
-    public Usuario actualizar(@PathVariable Long id, @RequestBody Usuario usuario) {
-        return usuarioService.update(id, usuario);
-    }
-
-    @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable Long id) {
-        usuarioService.delete(id);
-    }
-}
