--- biblioteca/src/main/java/Enums/EstadoLibro.java
@@ -1,6 +1,0 @@
-package Enums;
-
-public enum EstadoLibro {
-    DISPONIBLE,
-    PRESTADO
-}
