--- biblioteca/src/main/java/Enums/EstadoUsuario.java
@@ -1,6 +1,0 @@
-package Enums;
-
-public enum EstadoUsuario {
-    ACTIVO,
-    INACTIVO
-}
