--- biblioteca/src/main/java/Exceptions/PrestamoNoValidoException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class PrestamoNoValidoException extends RuntimeException {
-    public PrestamoNoValidoException(Long id) {
-        super("Prestamo con ID " + id + " no fue encontrado.");
-    }
-}
