--- biblioteca/src/main/java/Exceptions/UsuarioNoEncontradoException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class UsuarioNoEncontradoException extends RuntimeException {
-    public UsuarioNoEncontradoException(Long id) {
-        super("Usuario con ID " + id + " no fue encontrado.");
-    }
-}
