--- biblioteca/src/main/java/Modelos/Libro.java
@@ -1,20 +1,0 @@
-package Modelos;
-
-import Enums.EstadoLibro;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Libro {
-    private Long id;
-    private String isbn;
-    private String titulo;
-    private String autor;
-    private EstadoLibro estado;
-}
-
-
-
