--- biblioteca/src/main/java/Modelos/Prestamo.java
@@ -1,19 +1,0 @@
-package Modelos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Prestamo {
-    private Long id;
-    private Libro libro;
-    private Usuario usuario;
-    private LocalDate fechaPrestamo;
-    private LocalDate fechaDevolucion;
-
-}
