--- biblioteca/src/main/java/Services/LibroService.java
@@ -1,14 +1,0 @@
-package Services;
-
-import Modelos.Libro;
-
-import java.util.List;
-
-public interface LibroService {
-    Libro save(Libro libro);
-    Libro findByIsbn(String isbn);
-    Libro findById(Long id);
-    List<Libro> findAll();
-    void delete(Long id);
-    Libro update(Long id, Libro libro);
-}
