--- biblioteca/src/main/java/Services/UsuarioService.java
@@ -1,13 +1,0 @@
-package Services;
-
-import Modelos.Usuario;
-
-import java.util.List;
-
-public interface UsuarioService {
-    Usuario save(Usuario usuario);
-    Usuario findById(Long id);
-    List<Usuario> findAll();
-    void delete(Long id);
-    Usuario update(Long id, Usuario usuario);
-}
