+++ biblioteca/src/main/java/um/programacion2/biblioteca/controllers/LibroController.java
@@ -1,0 +1,46 @@
+package um.programacion2.biblioteca.controllers;
+
+import org.springframework.http.HttpStatus;
+import um.programacion2.biblioteca.modelos.Libro;
+import um.programacion2.biblioteca.services.LibroService;
+
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("api/libros")
+public class LibroController {
+    private final LibroService libroService;
+
+    public LibroController(LibroService libroService) {
+        this.libroService = libroService;
+    }
+
+    @GetMapping
+    public List<Libro> obtenerTodos() {
+        return libroService.findAll();
+    }
+
+    @GetMapping("/{id}")
+    public Libro obtenerPorId(@PathVariable Long id) {
+        return libroService.findById(id);
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public Libro crear(@RequestBody Libro libro) {
+        return libroService.save(libro);
+    }
+
+    @PutMapping("/{id}")
+    public Libro actualizar(@PathVariable Long id, @RequestBody Libro libro) {
+        return libroService.update(id, libro);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT) //
+    public void eliminar(@PathVariable Long id) {
+        libroService.delete(id);
+    }
+}
