+++ biblioteca/src/main/java/um/programacion2/biblioteca/enums/EstadoLibro.java
@@ -1,0 +1,6 @@
+package um.programacion2.biblioteca.enums;
+
+public enum EstadoLibro {
+    DISPONIBLE,
+    PRESTADO
+}
