+++ biblioteca/src/main/java/um/programacion2/biblioteca/exceptions/LibroNoEncontradoException.java
@@ -1,0 +1,12 @@
+package um.programacion2.biblioteca.exceptions;
+
+public class LibroNoEncontradoException extends RuntimeException {
+
+    public LibroNoEncontradoException(Long id) {
+        super("Libro con ID " + id + " no fue encontrado.");
+    }
+
+    public LibroNoEncontradoException(String isbn) {
+        super("Libro con ISBN " + isbn + " no fue encontrado.");
+    }
+}
