+++ biblioteca/src/main/java/um/programacion2/biblioteca/exceptions/PrestamoNoValidoException.java
@@ -1,0 +1,7 @@
+package um.programacion2.biblioteca.exceptions;
+
+public class PrestamoNoValidoException extends RuntimeException {
+    public PrestamoNoValidoException(Long id) {
+        super("Prestamo con ID " + id + " no fue encontrado.");
+    }
+}
