+++ biblioteca/src/main/java/um/programacion2/biblioteca/modelos/Prestamo.java
@@ -1,0 +1,19 @@
+package um.programacion2.biblioteca.modelos;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Prestamo {
+    private Long id;
+    private Libro libro;
+    private Usuario usuario;
+    private LocalDate fechaPrestamo;
+    private LocalDate fechaDevolucion;
+
+}
