+++ biblioteca/src/main/java/um/programacion2/biblioteca/services/LibroServiceImpl.java
@@ -1,0 +1,53 @@
+package um.programacion2.biblioteca.services;
+
+import um.programacion2.biblioteca.exceptions.LibroNoEncontradoException;
+import um.programacion2.biblioteca.modelos.Libro;
+import um.programacion2.biblioteca.repositories.LibroRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class LibroServiceImpl implements LibroService {
+    private final LibroRepository libroRepository;
+
+    public LibroServiceImpl(LibroRepository libroRepository) {
+        this.libroRepository = libroRepository;
+    }
+
+    @Override
+    public Libro save(Libro libro) {
+        return libroRepository.save(libro);
+    }
+
+    @Override
+    public Libro findByIsbn(String isbn) {
+        return libroRepository.findByIsbn(isbn)
+                .orElseThrow(() -> new LibroNoEncontradoException(isbn));
+    }
+    @Override
+    public Libro findById(Long id) {
+        return libroRepository.findById(id)
+                .orElseThrow(() -> new LibroNoEncontradoException(id));
+    }
+
+    @Override
+    public List<Libro> findAll() {
+        return libroRepository.findAll();
+    }
+
+    @Override
+    public void delete(Long id) {
+        libroRepository.deleteById(id);
+    }
+
+    @Override
+    public Libro update(Long id, Libro libro) {
+        if (!libroRepository.existsById(id)) {
+            throw new LibroNoEncontradoException(id);
+        }
+        libro.setId(id);
+        return libroRepository.save(libro);
+    }
+
+}
