+++ biblioteca/src/main/java/um/programacion2/biblioteca/services/UsuarioServiceImpl.java
@@ -1,0 +1,48 @@
+package um.programacion2.biblioteca.services;
+
+import um.programacion2.biblioteca.exceptions.UsuarioNoEncontradoException;
+import um.programacion2.biblioteca.modelos.Usuario;
+import um.programacion2.biblioteca.repositories.UsuarioRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class UsuarioServiceImpl implements UsuarioService {
+
+    private final UsuarioRepository usuarioRepository;
+
+    public UsuarioServiceImpl(UsuarioRepository usuarioRepository) {
+        this.usuarioRepository = usuarioRepository;
+    }
+
+    @Override
+    public Usuario save(Usuario usuario) {
+        return usuarioRepository.save(usuario);
+    }
+
+    @Override
+    public Usuario findById(Long id) {
+        return usuarioRepository.findById(id)
+                .orElseThrow(() -> new UsuarioNoEncontradoException(id));
+    }
+
+    @Override
+    public List<Usuario> findAll() {
+        return usuarioRepository.findAll();
+    }
+
+    @Override
+    public void delete(Long id) {
+        usuarioRepository.deleteById(id);
+    }
+
+    @Override
+    public Usuario update(Long id, Usuario usuario) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new UsuarioNoEncontradoException(id);
+        }
+        usuario.setId(id);
+        return usuarioRepository.save(usuario);
+    }
+}
